+++ pakyow-presenter/CHANGELOG.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+  * [fixed] Incorrect renders when a collapsed string doc is transformed
+
 # 1.0
 
   * Hello, Web